+++ libs/web/src/components/common/index.ts
@@ -28,3 +28,5 @@ export {SkeletonLoader} from "./skeleton-loader"
 export {CoinsListModal} from "./coins-list-modal"
 export {SwapSuccessModal} from "./swap-success-modal"
 export {CoinListItem} from "./coin-list-item"
+export {HeaderNew} from "./header-new"
+export {LogoNew} from "./logo-new"
